code-strategos/pyproject.toml
[project]
name = "code-strategos"
version = "0.1.0"
description = "A multi-agent strategic code generation system powered by CrewAI"
dependencies = [
    "crewai>=0.121.0",
    "crewai-tools>=0.45.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.45.0",
]
requires-python = ">=3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
